# Speed up datatype transformer