This speeds up the `DataTypeTransformer` by roughly 30%, based on profiling a very slow RealtimeToOfflineTask mapping phase in #8014, where over 30% of samples from a 5 minute profile were in this method.

before
```
Benchmark                               (_columnCount)  Mode  Cnt     Score    Error  Units
BenchmarkDataTypeTransformer.transform              10  avgt    5   487.921 ± 23.518  ns/op
BenchmarkDataTypeTransformer.transform              20  avgt    5  1026.615 ± 21.386  ns/op
BenchmarkDataTypeTransformer.transform              30  avgt    5  1540.456 ± 23.694  ns/op
BenchmarkDataTypeTransformer.transform              40  avgt    5  2086.853 ± 84.693  ns/op
BenchmarkDataTypeTransformer.transform              50  avgt    5  2649.823 ± 65.988  ns/op
```
after
```
Benchmark                               (_columnCount)  Mode  Cnt     Score     Error  Units
BenchmarkDataTypeTransformer.transform              10  avgt    5   386.467 ±   6.199  ns/op
BenchmarkDataTypeTransformer.transform              20  avgt    5   781.351 ±  30.718  ns/op
BenchmarkDataTypeTransformer.transform              30  avgt    5  1148.439 ±  52.494  ns/op
BenchmarkDataTypeTransformer.transform              40  avgt    5  1635.082 ±  97.129  ns/op
BenchmarkDataTypeTransformer.transform              50  avgt    5  2039.217 ± 100.647  ns/op
```