@richardstartin 2022-01-13
When the message is large and the number of field specs is small, this will not be beneficial